+++ src/cli/commands/public-cmds/use-cmd.js
@@ -1,15 +1,12 @@
 /** @flow */
 import chalk from 'chalk';
 import Command from '../../command';
-import { use } from '../../../api/consumer';
-import type {
-  UseProps,
-  MergeStrategy,
-  SwitchVersionResults,
-  ApplyVersionResult
-} from '../../../consumer/component/switch-version';
-import { MergeOptions } from '../../../consumer/component/switch-version';
 import { BitId } from '../../../bit-id';
+import { use } from '../../../api/consumer';
+import { applyVersionReport } from './merge-cmd';
+import { getMergeStrategy } from '../../../consumer/versions-ops/merge-version';
+import type { UseProps } from '../../../consumer/versions-ops/checkout-version';
+import type { ApplyVersionResults } from '../../../consumer/versions-ops/merge-version';
 
 export default class Use extends Command {
   name = 'use <version> <ids...>';
@@ -30,7 +27,7 @@ export default class Use extends Command {
     ['M', 'manual', 'in case of a conflict, leave the files with a conflict state to resolve them manually later'],
     ['v', 'verbose', 'showing verbose output for inspection'],
     ['', 'skip-npm-install', 'do not install packages of the imported components'],
-    ['', 'ignore-dist', 'write dist files (when exist) to the configured directory']
+    ['', 'ignore-dist', 'do not write dist files (when exist)']
   ];
   loader = true;
 
@@ -53,23 +50,13 @@ export default class Use extends Command {
       skipNpmInstall?: boolean,
       ignoreDist?: boolean
     }
-  ): Promise<SwitchVersionResults> {
-    const getMergeStrategy = (): ?MergeStrategy => {
-      if ((ours && theirs) || (ours && manual) || (theirs && manual)) {
-        throw new Error('please choose only one of the following: ours, theirs or manual');
-      }
-      if (ours) return MergeOptions.ours;
-      if (theirs) return MergeOptions.theirs;
-      if (manual) return MergeOptions.manual;
-      return null;
-    };
-
+  ): Promise<ApplyVersionResults> {
     const bitIds = ids.map(id => BitId.parse(id));
     const useProps: UseProps = {
       version,
       ids: bitIds,
       promptMergeOptions: merge,
-      mergeStrategy: getMergeStrategy(),
+      mergeStrategy: getMergeStrategy(ours, theirs, manual),
       verbose,
       skipNpmInstall,
       ignoreDist
@@ -77,17 +64,9 @@ export default class Use extends Command {
     return use(useProps);
   }
 
-  report({ components, version }: SwitchVersionResults): string {
+  report({ components, version }: ApplyVersionResults): string {
     const title = `the following components were switched to version ${chalk.bold(version)}\n`;
-    const componentsStr = components
-      .map((component: ApplyVersionResult) => {
-        const name = component.id.toStringWithoutVersion();
-        const files = Object.keys(component.filesStatus)
-          .map(file => `\t${chalk.bold(file)} => ${component.filesStatus[file]}`)
-          .join('\n');
-        return `${name}\n${chalk.cyan(files)}`;
-      })
-      .join('\n\n');
+    const componentsStr = applyVersionReport(components);
     return chalk.underline(title) + chalk.green(componentsStr);
   }
 }